+++ ui/mods/com.pa.quitch.galactic-war-multiplanetary-systems/shared/build_map_list.js
@@ -7,21 +7,18 @@ if (!multiplanetarySystemTabsLoaded) {
     try {
       var multiplanetaryMaps = [];
       var multiStartMaps = [];
+      var singlePlanetMaps = [];
       var mapTabOne = loc("!LOC:Multiplanetary Systems");
       var mapTabTwo = loc("!LOC:Multiplanetary Spawns");
+      var mapTabThree = loc("!LOC:Single Planet Systems");
 
       // Create an empty tab to load in time for Shared Systems for Galactic War
       cShareSystems.load_pas(mapTabOne, multiplanetaryMaps);
       cShareSystems.load_pas(mapTabTwo, multiStartMaps);
+      cShareSystems.load_pas(mapTabThree, singlePlanetMaps);
 
-      var checkForMultiplePlanets = function (
-        numberOfPlanets,
-        filePathOrSystem
-      ) {
-        if (numberOfPlanets > 1) {
-          return filePathOrSystem;
-        }
-        return undefined;
+      var checkForMultiplePlanets = function (numberOfPlanets) {
+        return numberOfPlanets > 1;
       };
 
       var checkForMultiplanetarySpawns = function (planets, filePathOrSystem) {
@@ -41,12 +38,17 @@ if (!multiplanetarySystemTabsLoaded) {
         planets,
         multiPlanetMaps,
         multiSpawnMaps,
+        singlePlanets,
         filePathOrSystem
       ) {
-        var multiPlanetResult = checkForMultiplePlanets(
-          planets.length,
-          filePathOrSystem
-        );
+        var multiPlanetResult; // String or Object
+        var singlePlanetResult; // String or Object
+
+        if (checkForMultiplePlanets(planets.length) === true) {
+          multiPlanetResult = filePathOrSystem;
+        } else {
+          singlePlanetResult = filePathOrSystem;
+        }
         var multiStartResult = checkForMultiplanetarySpawns(
           planets,
           filePathOrSystem
@@ -58,18 +60,23 @@ if (!multiplanetarySystemTabsLoaded) {
         if (!_.isUndefined(multiStartResult)) {
           multiSpawnMaps.push(multiStartResult);
         }
+        if (!_.isUndefined(singlePlanetResult)) {
+          singlePlanets.push(singlePlanetResult);
+        }
       };
 
       var processDefaultSystems = function (
         systems,
         multiPlanetMaps,
-        multiSpawnMaps
+        multiSpawnMaps,
+        singlePlanets
       ) {
         _.forEach(systems, function (system) {
           processSystems(
             system.planets,
             multiPlanetMaps,
             multiSpawnMaps,
+            singlePlanets,
             system
           );
         });
@@ -80,6 +87,7 @@ if (!multiplanetarySystemTabsLoaded) {
       ) {
         var defaultMultiplanetary = [];
         var defaultMultiStart = [];
+        var defaultSingleSystem = [];
 
         // Protect against failure in gw_start
         if (!_.isUndefined(model.cShareSystems_tabsIndex)) {
@@ -92,22 +100,26 @@ if (!multiplanetarySystemTabsLoaded) {
             processDefaultSystems(
               systems,
               defaultMultiplanetary,
-              defaultMultiStart
+              defaultMultiStart,
+              defaultSingleSystem
             );
           });
           processDefaultSystems(
             premadeSystems,
             defaultMultiplanetary,
-            defaultMultiStart
+            defaultMultiStart,
+            defaultSingleSystem
           );
 
           // Add My Systems and PA maps when tabs are ready
           var addedDefaultMultiSystems = false;
           var addedDefaultMultiStarts = false;
+          var addedDefaultSingleSystem = false;
           model.cShareSystems_tabsIndex.subscribe(function (tabs) {
             if (
               addedDefaultMultiSystems === false ||
-              addedDefaultMultiStarts === false
+              addedDefaultMultiStarts === false ||
+              addedDefaultSingleSystem === false
             ) {
               _.forEach(tabs, function (tab) {
                 if (
@@ -122,6 +134,12 @@ if (!multiplanetarySystemTabsLoaded) {
                 ) {
                   tab.systems(tab.systems().concat(defaultMultiStart));
                   addedDefaultMultiStarts = true;
+                } else if (
+                  tab.name === mapTabThree &&
+                  addedDefaultSingleSystem === false
+                ) {
+                  tab.systems(tab.systems().concat(defaultSingleSystem));
+                  addedDefaultSingleSystem = true;
                 }
               });
             }
@@ -156,6 +174,7 @@ if (!multiplanetarySystemTabsLoaded) {
                 mapFile.planets,
                 multiplanetaryMaps,
                 multiStartMaps,
+                singlePlanetMaps,
                 coherentFilePath
               );
             }).always(function () {
@@ -166,12 +185,14 @@ if (!multiplanetarySystemTabsLoaded) {
           $.when.apply($, deferredQueue).then(function () {
             cShareSystems.load_pas(mapTabOne, multiplanetaryMaps);
             cShareSystems.load_pas(mapTabTwo, multiStartMaps);
+            cShareSystems.load_pas(mapTabThree, singlePlanetMaps);
           });
 
           // Fallback for using the mod without map packs
           if (noMapPacksInstalled === true) {
             cShareSystems.addTab(mapTabOne, defaultMultiplanetary);
             cShareSystems.addTab(mapTabTwo, defaultMultiStart);
+            cShareSystems.addTab(mapTabThree, singlePlanetMaps);
           }
         });
       });